requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pymilvus==2.3.4
milvus==2.3.4
openai>=1.12.0
pypdf==3.17.4
python-dotenv==1.0.0
pydantic==2.5.0
python-multipart==0.0.6
numpy>=1.24.0
marshmallow>=3.13.0
environs>=10.0.0

